Remove unnecessary BCOffsetBase::operator>
Summary:
There's already operator<=>, but that needed more flexible argument types to
generate the <, >, <=, >= comparison operators it seems.

I can't explain why the == operators still need to be here.  Still something to
do with how the <=> operators are defined, but I can't make heads or tails of
it.  Either way this is an improvement as we get to drop >.

Reviewed By: swtaarrs

Differential Revision: D41347059

fbshipit-source-id: 451892e241a83775ad7da8c0d741390ed30ed49d