Delete LIR rewrite optimizations for `CondBranch`
Summary:
`jit::lir::PostRegAllocRewrite::doRewriteCondBranch` is responsible
for lowering `CondBranch` LIR instructions into a test/jcc. In
certain cases it would omit the test if it was deemed redundant.
In at least one case, the logic for determining whether or not the
test was needed was based on unsafe assumptions. I suspect the other
cases may be making faulty assumptions as well. Rather than audit
and fix each case, which is easy to get wrong, we instead always
emit a test/jcc. This is less efficient but should never be incorrect.
We can add back correct optimizations if it turns out they're actually
needed for performance.

`jit::lir::PostRegAllocRewrite::doRewriteCondBranch` will omit the test
if the flag affecting instruction that precedes the `CondBranch` being
written is a compare. However, it did not check whether or not the compare
was testing the operand of the `CondBranch`. `doRewriteCondBranch` would
erroneously omit the test when the preceding flag affecting instruction
was a compare that did not test the operand of the `CondBranch`. For example,

```
        RAX:8bit = Equal RAX:Object, 13010792(0xc68768):Object

[RDI:Object + 0x18]:8bit = Move RAX:8bit

      RAX:Object = Move 140588694798896(0x7fdd5b368230):Object

      RBX:Object = Move [RDI:Object + 0x10]:Object

[RDI:Object + 0x10]:Object = Move RAX:Object

                   CondBranch RBX:Object
```

would be transformed into the following:

```
        RAX:8bit = Equal RAX:Object, 13010792(0xc68768):Object

[RDI:Object + 0x18]:8bit = Move RAX:8bit

      RAX:Object = Move 140588694798896(0x7fdd5b368230):Object

      RBX:Object = Move [RDI:Object + 0x10]:Object

[RDI:Object + 0x10]:Object = Move RAX:Object

                   BranchZ BB%23
```

This happens because `doRewriteCondBranch` finds `Equal` as the flag affecting
instruction and decides to elide the test, however, the `Equal` does not test
`RBX`.

Additionally, `jit::lir::PostGenerationRewrite::rewriteCondBranch` implicitly
depended on `jit::lir::PostRegAllocRewrite::doRewriteCondBranch` for correctness,
so that was removed as well.

These fixes ended up uncovering a bug in the LIR we were using to implement
`Opcode::kIsNegativeAndErrOccurred'. The intent was to compare the contents
of a 32 bit register against -1. However, we were supplying the unsigned 32
bit representation of -1 when emitting LIR. For example,

```
%8:Object = NotEqual %7:32bit, 4294967295(0xffffffff):Object
```

Since the unsigned 32 bit represention of -1 doesn't fit into a signed
32 bit integer, the postgen rewrites inserted a move of the value into
a 64 bit register:

```
%9:Object = Move 4294967295(0xffffffff):Object
%8:Object = NotEqual %7:32bit, %9:Object
```

Code generation for this fails as we cannot emit a compare between a
32 bit and 64 bit register.

Reviewed By: carljm, itamaro

Differential Revision: D41418873

fbshipit-source-id: 523ecd9299c98632a611d44015d2cf16e6632d67