Include exception args in strict error message
Summary:
The display string of `StrictModuleUnhandledException`, which is displayed to the user, only includes the name of the exception, which doesn't always provide enough context to understand the root cause for the exception.
In some cases a more detailed message is included in the `exceptionArgs_`, but isn't displayed to the user.

This diff adds the `exceptionArgs_` to the displayed string, if they exist.

Reviewed By: czardoz

Differential Revision: D41519273

fbshipit-source-id: b568882cf3dfacf8ee9e29acb9927a4573b5589d