skip BuiltinLoadMethodElimination with duplicate CallMethod
Summary:
In Python 3.10, the compiler sometimes duplicates function exit blocks. In the
added test function, that means we get bytecode like this:

```
4634           0 LOAD_CONST               1 ('{}{}')
               2 LOAD_METHOD              0 (format)

4635           4 LOAD_CONST               2 (1)

4636           6 LOAD_FAST                1 (flag)
               8 POP_JUMP_IF_TRUE         8 (to 16)
              10 LOAD_CONST               3 ('')

4634          12 CALL_METHOD              2
              14 RETURN_VALUE

4636     >>   16 LOAD_CONST               4 (' flag')

4634          18 CALL_METHOD              2
              20 RETURN_VALUE
```

where we have two different `CALL_METHOD` both using the output of the same
`LOAD_METHOD`. This means we end up with jit HIR with two different
`CallMethod` in different exit blocks using the output of the same `LoadMethod`.

Our `BuiltinLoadMethodElimination` optimization pass wasn't built with this
possibility in mind; it assumes a 1:1 correspondence between `LoadMethod`
and `CallMethod`.

It may be possible to refactor the optimization pass so it can handle this (and
replace multiple `CallMethod` instructions while replacing a single `LoadMethod`
and `GetLoadMethodInstance`), but for now I just chose to bail on the
optimization entirely if we detect this situation.

Reviewed By: itamaro, tekknolagi

Differential Revision: D41489047

fbshipit-source-id: a034de7be51f7232428c945ed3249198cd71bb1c