don't inline comprehensions that might have cells
Summary:
Port D40819342.

Original summary:

Our strategy for safely inlining comprehensions is to delete any names
that it defines right after running the inlined comprehension code, so it
doesn't have a visible effect on the scope we inlined it into (and we refuse
to inline if there are any colliding names already in the outer scope.)

Currently we will inline a comprehension that defines a name which is also a
cell (i.e. it's closed over by an inner scope.) But this is not safe; unlike a
local var, the value in the cell is relevant beyond just while the comprehension
code is running, and if another inlined comprehension later in the function
overwrites that value, it will visibly change semantics. E.g. see the added test
case `test_do_not_inline_comp_with_cells`.

In `symtable.c` in order to determine if a free var in a nested scope is a cell
or just e.g. a reference to a global, we'd have to recursively walk all child
scopes and collect all free var names, then see if any are defined in the
comprehension. Instead of doing all that, we just check the existing
`ste_child_free` metadata and refuse to inline if the comprehension has any
contained scope with free vars. This should be fine, as lambdas inside
comprehensions are quite rare, let alone ones with free vars.

Reviewed By: itamaro

Differential Revision: D41486754

fbshipit-source-id: 8112b0c4ac5ee20ed12da2484c135952484365d1