don't try to get a globals-cache when builtins/globals are unwatchable
Summary:
Port D40948446.

Original summary:

The way our GlobalCache is built, it assumes that we won't even try to create
one if the relevant globals and builtins are not watchable. Maybe it should not
assume this, but it will require some additional refactoring to allow it to signal
that "can't cache" state to its callers. In the meantime, it's the responsibility of
the caller (which in practice is always the preloader) to not call `findGlobalCache`
unless we know the builtins and globals are watchable.

We check this in `preload()` before adding a name to the list of cachable globals,
but it's possible for something later in preloading to make builtins or globals
unwatchable, and then we fail later in `global()` when we're actually ready to emit
code for that `LOAD_GLOBAL` instruction.

We actually had a test for this case already
(`test_preload_side_effect_makes_globals_unwatchable`) which failed if run in
isolation but unfortunately passed for bad reasons if it was run after another
test (`test_preload_side_effect_makes_builtins_unwatchable`), because the
latter test made builtins unwatchable which in effect bypassed the entire
meat of the former test (since we'd never try to cache any globals at all.)

Reviewed By: itamaro, tekknolagi

Differential Revision: D41487105

fbshipit-source-id: 98ea109e7582ac06478243478e95db41cee30d41