prefetch global value in preloader so GlobalCache won't self-invalidate
Summary:
Port of D41351393 from `cinder/3.8`, minus the revert of D40956578,
which I never ported to 3.10 in the first place.

Original summary, slightly edited to remove references to reverting
D40956578:

This diff works around the problem of a GlobalCache invalidating itself
mid-init due to lazy import side effects. This would occur only the
first time we created a globals cache, in `Preloader::preload`.  Before
trying to create the GlobalCache, we can first pre-fetch the value from
globals/builtins dicts. This ensures that by the time we actually create
a GlobalCache, we aren't susceptible to side effects anymore.

This is unsatisfying, as it means the implementation of GlobalCache
remains unsafe in the presence of lazy imports; it relies on its caller
ensuring any possible lazy imports have been resolved before creating the
GlobalCache. Technically this means that shadowcode's use of
GlobalCache could still be susceptible, but in practice it isn't because
shadowcode doesn't kick in until a given code object has executed 50
times, by which time any lazy imports loaded by globals will likely have
been resolved. (It may be possible to construct a synthetic repro where a
global referring to a lazy import is only referenced in a code path which
executes only on the 51st+ execution of a function; I'll look into this and
put a mitigation in place if I can reproduce it.)

Alternatives that I wasn't able to make work or decided not to pursue for
now:

1) I tried having the `Runtime` hold the reference to the name, instead of
the GlobalCache holding it, but this doesn't solve the refleak problem;
there are too many new GlobalCaches created for various reasons in Nth
runs of a given test, and each one can still leak a new name.

2) I tried making GlobalCache objects be stateful (with a `disabled_` bool)
instead of just stateless wrappers around a `GlobalCacheMap::value_type`,
but this doesn't work because GlobalCache objects are passed by value;
I didn't want to create the need to start managing their lifetime, too.

3) I tried having `GlobalCache::disable` null out the `name` pointer on the
wrapped `GlobalCacheKey` so as to release the reference, but
`GlobalCacheMap::value_type` is a pair whose first element is a
`const GlobalCacheKey`. This would require deeper changes to work
around and I suspected that trying to in-place mutate the key in a map
would cause problems I didn't want to deal with.

4) I tried moving the `PyDict_GetItem` out of `GlobalCache::init` and into
the caller (`Runtime::findGlobalCache`). This gets complicated
when considering all the constraints: non-initial `findGlobalCache` should
be low overhead and not do extra dict fetches, JIT should never do dict
fetches during multi-threaded compilation, callers of `findGlobalCache`
would need to be able to handle it failing, etc. I think this is still the right
direction for a more thorough fix, but wanted to get something up to
resolve the immediate issue first.

Reviewed By: tekknolagi

Differential Revision: D41563121

fbshipit-source-id: 81e60bcd19d0c63c01ca7802f36c84a84c579d63