Include exception args in strict error message
Summary:
The display string of `StrictModuleUnhandledException`, which is displayed to the user, only includes the name of the exception, which doesn't always provide enough context to understand the root cause for the exception.
In some cases a more detailed message is included in the `exceptionArgs_`, but isn't displayed to the user.

This diff adds the `exceptionArgs_` to the displayed string, if they exist.

Reviewed By: carljm

Differential Revision: D41546360

fbshipit-source-id: c3b96c08ab3fa189b08be8506f488debb675c903