Use structured API to generate LIR for CondBranch{,IterNotDone}
Summary:
This looks weird, because of how the CondBranch LIR instruction has no block
inputs.  That's because the blocks are all fixed up later in the end of
LIRGenerator::TranslateFunction().

This needed a fix to generate a Move for RefineType instructions.  The
getDefInstr() method was expecting to be able to load any pre-existing HIR
register as an LIR instruction (except for phi sources but this isn't a phi).
Turns out RefineType wasn't generating any LIR instructions, so its dest
register was not in the output map, fixing that with an explicit Move that can
get cleaned up later.

Reviewed By: swtaarrs

Differential Revision: D41354895

fbshipit-source-id: 0a7e61b0ac782ea9f9d026791f81917149c7f888