Update deferred compilation pointers as soon as target function is compiled
Summary:
During JIT compilation we sometimes want to insert an optimized JIT -> JIT static call for which the callee cannot immediately be recursively compiled. To work around this we inserted a "deferred" compilation call: `JITRT_CompileFunction()`. This tries to compile the target later as the caller executes and then patch itself away via an inline cache.

After investigating, I discovered practically deferred compilation only happens when: the target function can never be compiled (e.g. not on the JIT-list), or there was a race/recursion limit during multi-threaded batch compile. In the latter case the target function should eventually get compiled anyway as part of the rest of the batch. This means we can do away with deferred compilation in `JITRT_CompileFunction()`, and instead just update the inline cache pointer whenever the target function is naturally compiled.

Note we still need something like `JITRT_CompileFunction()` to handle the case where the callee actually can never compile but we don't find this until later. For this case I've renamed the function `JITRT_FailedDeferredCompileShim()` and it no longer attempts compilation.

There are three benefits to this approach:
1. We have no CoW in workers when they update their inline cache.
2. The way I've implemented this we can patch the inline cache with a call directly to the static entry point. Prior to this an indirection was needed which we stashed above the vectorcall entry point.
3. I believe this is simpler.

Reviewed By: carljm

Differential Revision: D36984428

fbshipit-source-id: 8ce181c1d8ae150274f4515f98b5244acdc971d0