Allow overloading recognized decorators
Summary:
Port of D40881327.

Original Summary:
Currently, the overload decorator doesn't allow overloads of
classmethods and staticmethods. This change gets them to be allowed.

Note: There's a different but related issue, where we don't allow
overloads for unrecognized decorators. That one's a little gnarly to
fix, because our policy for unknown decorators is that we don't process
any of them at all.

If we were to fix that, it'd perhaps involve trying to recognize the overload
decorator, in `Function().finish_bind()`, and setting a flag on
`UnknownDecoratedMethod`s. That's not a pressing matter though.

Reviewed By: itamaro

Differential Revision: D41658466

fbshipit-source-id: 58a2c573da5482e080dc1cdfff628e1d091c5894