Use the AST optimizer to get rid of dead code
Summary:
Port of D40911543.

Original Summary:
This small change helps get rid of dead code gated by
sys.hexversion checks.

We do not want to keep this around for long. But Static Python doesn't
do well when having multiple class definitions for the same class,
because it cannot determine which one will be used at runtime, so it
might as well just be dynamic.

In this case, we can safely recognize and remove some paths as dead
code.

Reviewed By: carljm

Differential Revision: D41562717

fbshipit-source-id: a4aff8ad38548b22c20b758fd44681c33796e789