Make sys.hexversion checking more robust
Summary:
Sometimes, a given static module will need to support multiple Python
versions. As a short term hack, D40911543 implemented an AST rewrite that
removes the "dead" code from the AST altogether. This is quite a hacky
approach, because removing the nodes is a destructive action done at a really
early stage. It also creates incompatibilities between the native (written in
C) compiler and the Python compiler.

In this diff, we introduce a more structured approach to the implementation,
where we skip visiting the dead code in decl-visit, type-checking and code-gen
passes.

Reviewed By: carljm

Differential Revision:
D41596503

LaMa Project: L1125187

fbshipit-source-id: 34ce26562dd65936ea1dada60aecd695e56f5e7c