remove obsolete python38.sh script
Summary:
This looks like it was added around the same time as D25855938 as part of
creating a "baseline CPython 3.8" fbpkg for running microbenchmarks. But
no versions of that fbpkg (`python.cinder.cpython38`) exist, and nothing else
uses this script.

This script ends up getting flagged as a "use of platform009" due to the path.

Reviewed By: tekknolagi

Differential Revision: D41667249

fbshipit-source-id: c0bf0508140035938b7cb0a8498ff7b0ddf68fe8