Make sys.hexversion checking more robust
Summary:
Port of D41596503

The only significant change from that diff is, we override
`CodeGenerator.visitIf` in `Static38CodeGenerator`. The reason for this
is, we can't just rely on `maybeEmit` to stop new bytecode generation,
we need to avoid visiting these AST nodes altogether.

Original commit message:

Sometimes, a given static module will need to support multiple Python
versions. As a short term hack, D40911543 implemented an AST rewrite that
removes the "dead" code from the AST altogether. This is quite a hacky
approach, because removing the nodes is a destructive action done at a really
early stage. It also creates incompatibilities between the native (written in
C) compiler and the Python compiler.

In this diff, we introduce a more structured approach to the implementation,
where we skip visiting the dead code in decl-visit, type-checking and code-gen
passes.

Reviewed By: carljm

Differential Revision:
D41659871

LaMa Project: L1125187

fbshipit-source-id: e91476756cd27c1c80139e304029ad932d21e278