Use materialized Python frame when restoring TOS in `_gen_throw`
Summary:
3.10 added code in `_gen_throw` that, when throwing into a generator
that is yielding from another value, does the following:

1. Saves the top of the Python stack to a local.
2. Sets the top of the Python stack to the frame for the generator.
3. Throws into the yielded-from value.
4. Restores the top of the Python stack to the value from (1).

We extended this logic in cinder to save/restore the top of the shadow
stack as well. However, the top of the shadow stack may not have a Python
frame associated with it during (1) if we're running in shadow frame
mode. During (3), the Python frame for the top of the shadow stack
that was saved in (1) may be materialized. When that happens, we need to
restore the top of the Python stack to the frame that was materialized
in (3).

Reviewed By: carljm

Differential Revision: D41753574

fbshipit-source-id: 6c14eb4be9acb58d62d64862f31cfa0d29530491