materialize shadow call stack even after jit is disabled
Summary:
In Python 3.8, there was a convenient "frame-getter" hook that we used to
ensure we would materialize the shadow-frame stack anytime a Python frame
was requested. We would install this frame getter whenever the JIT was
initialized, and never uninstall it:
https://github.com/facebookincubator/cinder/blob/cinder/3.8/Jit/pyjit.cpp#L1448-L1449

In 3.10 this hook disappeared, so instead we patch core functions like
`PyEval_GetFrame` and `PyThreadState_GetFrame` so that they defer to
`_PyJIT_GetFrame`, which materializes the shadow-frame stack only if the JIT
is enabled:
https://github.com/facebookincubator/cinder/blob/cinder/3.10/Jit/pyjit.cpp#L2346-L2351

This introduced a subtle breakage: if the JIT is initially enabled, and we compile
a bunch of functions, and then we disable the JIT, thereafter the JIT compiled
functions will still run (and use shadow-frames), but we will wrongly fail to
materialize the shadow-frame stack when Python frames are needed.

Unfortunately the production environment disables the JIT before fork to avoid
JIT compilation in worker processes, which triggers this bug.

Instead we should match the 3.8 behavior by materializing shadow-frames if
the JIT has ever been initialized, even if it has since been disabled.

Reviewed By: mpage

Differential Revision: D41787433

fbshipit-source-id: 1043044f2d984db8cdf18f76f7c85ef21b536714