Remove read-only features
Summary:
Like some sort of philistine code-butcher, in this diff I've gone and removed all the "read-only" features from Cinder. Unfortunately, this experimental project has come to an end and this is all dead code.

I didn't know a huge amount about this features so I did this with the help of porting asserts, grepping for "readonly", and `git blame`.

Reviewed By: alexmalyshev

Differential Revision: D42121399

fbshipit-source-id: 3d9e861df4d09dc60ed4b1075eaf0ab7acefd007