Capture current context in 'ContextAwareTask.add_done_callback'
Summary:
Port of D26395878 from cinder/3.8.

The lack of this `add_done_callback` override was the cause of the
linked task.

Original summary:
This diff adds override for `add_done_callback` to `ContextAwareTask` so callback will run with the same context that was captured in the task at the point when `add_done_callback` was invoked.

Note: asyncio module uses `add_done_callback` internally in two cases:
- adding completion callback for arguments of `gather` to record the result in `GatheringFuture`
- adding completion callback to future that caused task to be suspended to awake the task when future is fulfilled.

We still use original `add_done_callback` implementation from the `Future` in this scenarios since we know exactly what callbacks supposed to do and it does not need context to be  set. To emphasize that method tables don't necessary use exact `add_done_callback` for the type I've renamed the slot to `on_completed` to avoid possible confusion in the future.

Reviewed By: alexmalyshev

Differential Revision: D42158735

fbshipit-source-id: ec5db6a8af9ff6078c82e153eb310c3f3d36cae3