Default Imm to OperandBase::k64bit
Summary:
Currently Imm will default to kObject, but in reality we're generally using it
for raw integral values.  We should be using MemImm for any kObject uses going
forward.

There's a bit of awkwardness here with `LoadConst<Nullptr>` being the same as
`LoadConst<0>` as far as LIRGenerator is concerned.  I feel like `Nullptr`
should continue to get typed as kObject, so this is still explicitly using
kObject for that case.

Reviewed By: swtaarrs

Differential Revision: D42072559

fbshipit-source-id: 5b706731ef8a80da4eb71cb3f54caaa9b7bbd73c