Simplify LoadVarObjectSize with known tuple to LoadConst
Summary:
This ports D41357554 from Cinder 3.8.  There is a Cinder 3.10 version of this
diff lying around (D40327265), but it's older, so I chose to port this one.
The summary below this line is all copied.

We use this in situations where we want to check that a bunch
of loads will succeed at once, like in UNPACK_SEQUENCE:

  def test():
    a, b, c = 1, 2, 3
    return a

Since we know the size of (1, 2, 3) at JIT time (it's a compile-time
constant), we can load it right off the PyObject*.

Reviewed By: brittanyrey

Differential Revision: D42350845

fbshipit-source-id: a09045bdc351f7d12fc0c37c810fff85894d56d4