Type the return of InvokeStaticFunction in LIR
Summary:
This ended up biting me while I was working on de-stringifying IntBinaryOp.  I
was getting a crash in `make testcinder_jit` due to an invalid `imul bl, cl`
instruction being generated (8-bit imul has to use ax and al).

I tracked down the issue to this small repro, P580265190.  Note how IntBinaryOp
is getting a value that comes from an InvokeStaticFunction.  In HIR, the return
value is typed as CInt8, but the LIR generator generated `Mul RAX:Object,
RCX:8bit`.
This means the `rewriteByteMultiply` function in postalloc.cpp
(https://fburl.com/code/770swxgu) can't kick in for cases like this because the
first operand is not being correctly typed as an 8-bit argument.

Adding the return type to the Call generated for the InvokeStaticFunction fixes
the crash for me.

Reviewed By: carljm

Differential Revision: D42143037

fbshipit-source-id: 80b52b4cefb7f8df98da93278ccdc068e497bd8c