Structured LIR generation for IsNegativeAndErrOccurred
Summary:
This tweaks the API for creating new branching instructions.  The existing API
is odd in it works by block names and it automatically attaches successors if
the current block has <2 successors.  The newer form is to specify the branch
destinations as blocks which immediately sets them up as successors.  Right now
it's assuming there's always exactly 2 successors to a branch instruction.

Reviewed By: swtaarrs

Differential Revision: D42148607

fbshipit-source-id: 93a53801284f9ca9088589a6a9c7850595c0efae