Add str.format method definition to get a better error message
Summary:
The first step towards a full support for `str.format` in strict mode.

This change adds an "implementation" for the format method of `StrictString` that will always raise a `NotImplementedError` with a descriptive message. This is already an improvement from the current state, where calling `format` produces a `has no attribute` error.

Next step is to implement the method properly.

Reviewed By: carljm

Differential Revision: D42406044

fbshipit-source-id: c06e321dc8c8cdfaae0b51394debf92375db02d9