Use Operand::is*() functions instead of == OperandBase::k*
Summary:
LIR isn't consistent with how it inspects operand types.  Some of the code uses
helpers like `opnd->isImm()` but in other places we see `opnd->type() ==
OperandBase::kImm`.
This makes grepping for all the places where we use a specific type (e.g. immediates) harder.
So standardize on the is*() methods a little more.

Reviewed By: swtaarrs

Differential Revision: D42556796

fbshipit-source-id: e45898968402fed63c6fa498d9ee9199f141a099