Simplify PrimitiveCompare and PrimitiveBox
Summary:
This ports D41358454 from Cinder 3.8.  There is a Cinder 3.10 version of this
diff lying around (D39960215), but it's older so I chose to port this one.  I
had to merge the definitions of simplifyPrimitiveCompare between 3.8 and 3.10,
because they cover different cases and in one it's PrimitiveBox and in the other
it's PrimitiveBoxBool.

The summary below this line is all copied.

Backport of not-yet-landed D39960215.
Minimal constant propagation and code cleanup to clarify some
stuff for thatch.

Reviewed By: tekknolagi

Differential Revision: D42368523

fbshipit-source-id: 60dc69017c585009f1466b32dfa20b7dd808c095