Test that the expected type and number of deopts happen in profile tests
Summary:
`test_cinderjit_profile.py` is meant for end-to-end tests that ensure
that the specialization machinery works in the full context of normal managed
code, but it's easy accidentally deopt too early or for the wrong reason and
not test the specific codepath we care about.

This diff adds a context manager than can be used to express how many and what
kind of deopts we expect from certain blocks of code, and adds it to the
existing `BinaryOpTests`. The diffs stacked on top of this make further use of
it for `LOAD_ATTR` type specializations.

Since this file is now more narrowly targeted at exercising what happens in
type-specialized JIT code, I renamed it so the tests aren't run as part of the
normal `testcinder_jit` variants - the number and type of deopts will be
different, and none of the functionality tested is interesting under the
interpreter or non-profiled JIT.

Reviewed By: tekknolagi

Differential Revision: D37290209

fbshipit-source-id: ac137400cd284d17d7c05b7b5c7343ce7cf994da