Fix thread creation with interpreter type profiling enabled
Summary:
Interpreter type profiling is crashing in production under Cinder
3.10. The traceback pointed me at the call to
`Ci_ThreadState_SetProfileInterp()` from `new_threadstate()`, and I realized
that we were reading `tstate->cframe` before it was set by the caller. Move the
profiler initialization code to the end of the function, since we also want to
read `tstate->c_tracefunc` and `tstate->c_profilefunc`, which are set in
later in `new_threadstate()`.

Reviewed By: alexmalyshev

Differential Revision: D42584767

fbshipit-source-id: 56aa950f471a901dbe2c6c4a8fe625f32f359d86