str.format: check strictness for all args
Summary:
This diff adds support for any usage of `str.format` in strict mode.
We do that by converting all given arguments to string via their `__str__`/`__repr__` methods. If any of these calls is unsafe, a strict error is raised. This guarantees that the execution of `str.format` is strict.

This diff does not finish the task of adding support for `str.format` in strict modules, as it still doesn't return the correct value of the function. Right now the strict version of `format` will return a copy of the given string (the correct returned type). We do so simply because the full implementation of `format` is complex and requires more work.

Reviewed By: carljm

Differential Revision: D42665275

fbshipit-source-id: fd26d23e1c17884f35e599fcf008abe3f9454a4e