Use madvise for transparent huge pages rather than requesting them from mmap
Summary:
If we get huge pages from `mmap(..., MAP_HUGETLB)`, the process will die with
`SIGBUS` if we try to modify one of the resulting pages in a worker process
when there are no huge pages left for CoW. Too much CoW is bad for a host of
other reasons, and I have plans to separately address that, but if/when it does
happen in prod, we should gracefully degrade to small pages rather than
crashing.

This diff switches to allocating normal pages with `mmap()` then using
`madvise(..., MADV_HUGEPAGE)` to request transparent huge pages in the given
region.

Reviewed By: jbower-fb

Differential Revision: D42016867

fbshipit-source-id: f570d1e5105083efd4451d307caccf478df7b611