Immortal Instances 3.10 Port
Summary:
Immortalizing an object causes the instance to not participate in reference counting. The object is kept alive until runtime finalization, which helps us avoid unnecessary copy-on-writes for applications that share a Python heap across many processes.

This diff introduces a change to how the current immortal instances works by using the lower 32 bits of a 64 bit ref count and checking for overflow. To account for backwards compatibility since some extensions might modify the ref count instead we perform a memcpy on the lower 32 bits rather than simply assign an increment of the refcount to preserve this higher bit. Under this approach we decrease the number of instructions on INCREF and DECREF ad also provide an approach that will be backwards compatible with extensions built against a different python version.

In the official cpython the 4th bit will eventually be reserved and we can't use the existing method. This approach also gets us immortalization while mitigating the issue previously being hit where segfaults were occurring due to the way extensions were modifying the refcount

Reviewed By: eduardo-elizondo

Differential Revision: D42329724

fbshipit-source-id: 55be0efaf926e7cc22dd5d9614edf264f3daf415