Report error instead of crashing on invalid v-table ref
Summary: This just adds a slightly descriptive error message instead of crashing when we fail to find the v-table slot.  I've seen this error enough when things are broken that I thought we should improve this, but most recently I think whatever D42789360 (https://github.com/facebookincubator/cinder/commit/3e0e3784220d7d3a034e33850072a63c15e600ef) is hitting might be related to this as well. It seems when I had that in my Cinder build I was crashing out, so this might make it more obvious what's happening there.  But will be general debugging goodness for the future.

Reviewed By: czardoz

Differential Revision: D42981185

fbshipit-source-id: d8b1cdcaa7bef1c636da5fd423d8ea5514e1984a