Don't patch __build_class__, instead invoke it in static Python
Summary:
This removes our dependency upon patching `__build_class__`.  There is primarily one reason we currently need to patch this which is to fix-up descriptors w/ defaults for subclasses of static types defined in non-static Python.

This is replaced by automatically defining `__init_subclass__` on static Python types, or extending it when it's already defined.  We now cause `_static.init_subclass` to be invoked for the subtypes which can handle this fixup process.

For classes which are defined directly in static Python we'll still want some build class hooks going forward.  Therefore we replace the normal `LOAD_BUILD_CLASS` opcode with a direct invoke to `_static._build_cinder_class__`.  This gets a slightly different signature accepting only position args in the form of (function, name, kwargs dict) so that we can use `INVOKE_FUNCTION` to call it w/o keyword arguments.

Reviewed By: carljm

Differential Revision: D42790754

fbshipit-source-id: f123793aca1b28989d7861aca09b0cd1cddc0851