Moves static Python slot support out of typeobject.c
Summary:
This breaks apart static Python's dependency upon initialization of slots in typeobject.c.  This is done by using our custom `__build_cinder_class__` entry point for static classes where we will inspect `__slot_types__` and use it to fixup the type.

There's one significant limitation to this - if instances of the type have been created (via `__init_subclass__` or a metaclass overriding `__new__`) then this is potentially an unsafe operation.  We're potentially changing the size of the object, the locations where object references are held, and/or a place where an object reference might exist to a primitive value.  If any bad changes are going to happen we fail the creation of the type and report an error.  We attempt to detect if this has happened by using the reference count of the type and accounting for various references that get created during type creation.

This also requires us to re-do how our existing `__slots_with_default__` mechanism works.  The current implementation works by allowing the value to exist in the type and conflict with an entry in `__slots__`.  Because we need the slot to be created we instead now store the default value in the `__slots_with_default__` dictionary (which was previously just a tuple of names) and we remove it from the class.  When finishing up construction of the type we can create the descriptor with the default value.  In order to support that we also now need to track the overridden slots at runtime so they grow a new 'override' attribute.

After this there's probably a good amount of cleanup which can be done on our overall type construction.  We end up generating calls to multiple various helpers which can be probably be folded into our `__build_cinder_class__` helper, but that'll be saved for another diff.

Reviewed By: carljm

Differential Revision: D42876514

fbshipit-source-id: c6c5098fa57b743b7cd7608fdd8356f3a633778e