Make type code a clean extension
Summary: This pulls out our support for type codes from the custom data we add to types and moves it into the v-table.  It's pretty straight forward except for we need to re-create our v-table information when we clear our type cache information if we have a custom type code.

Reviewed By: carljm

Differential Revision: D42906408

fbshipit-source-id: 37e372d39864d4fb47a22af977dc9d33f58cd4dc