implement eval() to fix namedtuple stub
Summary:
In 3.10, namedtuple now uses `eval()` instead of `exec()`. When porting strict
modules analyzer to 3.10, we'd previously just hacked around that by copying
some code from the old namedtuple implementation into our stub. But this is still
breaking e.g. strict modules in some other codebases which don't have the
hacked-up stub.

I could copy the stub hack, but it seems nicer to implement `eval()` support and
move the stub back to using `implicit`, so I've done that here.

The core of the implementation is that we need to pass through a `Py_file_input`
or `Py_eval_input` down through to `_PyParser_ASTFromString` so that parsing
starts at the right part of the grammar (`eval()` only allows a single expression.)

In order to get nice errors if the code isn't valid (e.g. if we don't pass a single
expression to eval, but this also improves the existing error handling for exec
too) I added a `SyntaxError` type and a method to fetch the current Python
exception and raise it as a strict modules error.

The existing code also had support for an `Analyzer` with null `ModuleLoader`
(which would just raise an error if it hit an import). The analyzer for `exec()` used a
null `ModuleLoader`, and then it also checked for null `ModuleLoader` to prevent
nested exec/eval. But this caused problems because when a lambda is analyzed,
it gets the arena off the loader, which breaks if the loader is `nullptr`. It may be
possible to handle the arena differently here, but I also didn't see a reason why we
have to prevent import or nested exec/eval inside exec, so I just removed the
support for `nullptr` loader and always supply one.

Reviewed By: DinoV

Differential Revision: D43447110

fbshipit-source-id: f3e46a48193f92b5d68e52a11d6c7c85070cf6cc