Fix issue with reverse slot sorting based upon name
Summary:
Currently we sort the slots and reverse them, trying to turn something like `[(4, ('bar', )), (8, ('foo',))]` into `[(8, ('foo',)), (4, ('bar', ))]` so that the largest slots come first and smaller slots come later, maintaining the alignment for larger objects and minimizing the overall size of the object.  This was the previous logic and it worked just fine when creating the type was an atomic operation.

But when we have items of the same size this means they end up in reverse order by name.  Then when we try and re-initialize the slots after creation we end up with a layout difference and we report an error if we can't prove no instances of the type have been created.

This adds some custom size sorting logic which reverses the sorting only based upon size, not name of the fields.

Reviewed By: carljm

Differential Revision:
D43674471

Privacy Context Container: L1125187

fbshipit-source-id: bd26998a5cf20454c9181ad47d2cc1762e710e26