Add option to normalize filenames in profile data CodeKeys
Summary:
With a recent change, all paths to Python files in production include
a hash that changes with every deployment, whether or not the contents of the
file changed. We use the full filepath as part of the lookup key for code
objects and profile data is always from a few deployments ago, so this
effectively disabled type specialization.

Control for this by adding support for stripping a regex from filenames when
creating a `CodeKey`.

Reviewed By: mengdilin

Differential Revision:
D43542068

Privacy Context Container: L1125187

fbshipit-source-id: 2a7f20b488b828ecee15d11d970f253475a52048