Add LOAD_ATTR specialization for types with slots
Summary:
When the type of a `LoadAttr` receiver is a known, exact class with a
`T_OBJECT` or `T_OBJECT_EX` `PyMemberDescr`, we can compile it into a
`DeoptPatchPoint` plus a `LoadField`. This happens when a type is a Static
Python class or a normal Python class with `__slots__`.

Reviewed By: tekknolagi

Differential Revision:
D34014590

Privacy Context Container: L1125187

fbshipit-source-id: 28088d3db96735dfe3406a9caf0dd1b26f3d9afc