More precise deopt patching for slot-specialized LOAD_ATTR
Summary:
D34014590 by itself is a perf regression, because enough types change
after warmup that we deopt too much. Luckily, the parts of the types that
change aren't the parts we care about (for attribute slots, at least), so we
can deal with it by being more precise with the deopt patching.

This diff replaces `TypeDeoptPatcher` with `MemberDescrDeoptPatcher`, which
only patches the code when a type no longer has the specific slot we care
about. It also makes some minor tweaks to the general deopt patching mechanism
to support patchers that don't always patch.

Reviewed By: alexmalyshev

Differential Revision: D42512608

fbshipit-source-id: e2be35dc90ad50efa8379b903754b960294c4ee6