Mark all Cinder functions used by CinderVM
Summary:
This diff introduced `CiAPI_*` macros in `ci_api.h` for use whenever we add new symbols to core-Python code which are used by "CinderVM". The overall goal is to highlight when we are adding APIs which will likely make a pip installable module which includes all CinderVM features impossible.

Another way of looking at this is a simple grep for `CiAPI_` will highlight APIs which are at risk of breaking whenever we upgrade the CPython base for Cinder.

Reviewed By: itamaro

Differential Revision: D42701031

fbshipit-source-id: 583735583751ac27cdc85fd4912a872534b7c5d5