Make it possible to build `python` from Cinder codebase without CinderVM features
Summary:
This is a stepping stone to moving CinderVM into a separate `.so`. We add a `./configure` option `--disable-cinder-vm` to which allows us to build `python` without any features in CinderVM. This works by undef'ing a `EMBEDDED_CINDER` which disables any dependencies the core Python code now has on CinderVM.

Like the previous diff, this also explicitly exposes coupling we've introduced with core Python code which needs to be removed to make upgrades easier.

A material change in this diff is the `_strictmodule`, `_static`, `cinder`, and `xxclassloader` modules are now compiled as `.so` modules rather than linked statically into `libpython`. I've done this purely because this makes it easier to optionally exclude them from the build.

Reviewed By: itamaro

Differential Revision: D42761880

fbshipit-source-id: 00cc5085f7b77a492ee3f264925e71e01c805ed6