CinderVM .so build
Summary:
Another small step towards a CinderVM module, this diff introduces a CinderVM `.so` into the build. Note this isn't a "module" it's literally just taking things that may one day be modularized and putting them into an `.so` which is dynamically but explicitly loaded by `python`/`libpython.so`.

For now the contents of "CinderVM" are:
* All the C++ code from Cinder.
* The ThirdParty depdencies of this code.
* `classloader.c`.

This is all rather hacky and very much a proof-of-concept.

The new `.so` is built using distutils and has a `setup.py` in `Cinder/module/`. However, in order to run this `setup.py` we need a build of `python` from the same source tree. This is because distutils gets things like the C compiler to use, and build flags etc. from stuff baked into a Python distrubtion (IIUC).

So, the first step to building is to run `CInder/module/bootstrap.sh`. This builds a venv from the same source tree with just enough present to run `setup.py`. This generally only needs to be run once and the result lives in `build_cindervm_venv/` in the root of the repo.

Once the bootstrap venv exists, building CinderVM into an `.so` requires the `--enable-cinder-vm-so` flag being passed to `configure`, and then `make` as normal. The makefile shells out to `Cinder/module/build.sh` which wraps around `setup.py` to build the `.so`.

Reviewed By: itamaro

Differential Revision: D42867816

fbshipit-source-id: 7a51cfae48426f4cf528d6d3808535b4f3eb983f