Add test for regressions in CinderVM <-> core Python coupling
Summary:
Now we have the ability to build CinderVM as a `.so` we can build a `libpython.so` which would be used by the CinderVM `.so` and look at the `libpython.so` exported symbols. If these symbols change, it means we altered the API needed by the CinderVM `.so` (assuming the base CPython isn't changing). If we see new symbols we can detect this and suggest the user ensures they are exported with `CiAPI_` macros.

Generally the idea is to raise-the-bar a little for adding new couplings.

Reviewed By: itamaro

Differential Revision: D42908977

fbshipit-source-id: ec49d98eeee82f630a9d08d24ebe67ac0db11a5e