Make static type initialization more atomic
Summary:
We currently have a hodge-podge of initialization functions.  This moves that initialization logic into `__build_cinder_class__` by accepting more arguments.

With the logic spread out we have a hard time dealing with the case where v-tables are initialized eagerly because a base class has it's v-table initialized, forcing us to init the v-tables - and of course that initialization is invalid because we don't have all of the members populated yet.

Now we can force the re-initialization to happen after we've dealt with the cached properties because everything happens atomicly.

So we just grow a few new arguments to our `__build_cinder_class__`: the final method names, if the class is final, and the list of cached properties.  The list of cached properties has the prefixed name from which we can derive the normal name, and all the creation of the property objects is contained in the runtime logic now.

We keep around the existing `set_type_static` family because they are used for other non-codgen defined types.  But they are simplified to just set the bits.

There's also a small bug fix in here to actually decref the type on the failure case and some cleanups to use consts for the various arg offsets...  still kind of magical, but now with less repetition!

Reviewed By: carljm

Differential Revision: D44079029

fbshipit-source-id: e4ca18e2a2badffc6943bcf46d01cd46b6ef2dd1