Remove Ci_Py_VECTORCALL_INVOKED_CLASSMETHOD
Summary:
This eliminates the `Ci_Py_VECTORCALL_INVOKED_CLASSMETHOD` flag.  We now capture enough information in our lazy init vtable so that we can initialize the class method when we first execute it.  And then we get a new `type_vtable_coroutine_classmethod` function for handling the case when we actually have a class method in the slot for co-routines rather than relying on the flag.

From there the rest is pretty straight forward as we can just check to see if we're executing against a class method elsewhere.

This did expose one issue with class methods, and that is when we override them w/ a non-class method we are dropping the `self` on the overridden function.  This isn't super straight forward to resolve as we do the `LOAD_TYPE` in generated code to drop the instance.  For the time being I'm leaving the broken behavior as-is, but adding a test which notes it's incorrect.

Reviewed By: carljm

Differential Revision: D43747294

fbshipit-source-id: 2223d76d48d3b3214c29157fc8d0f83648f4ffdd